pyproject.toml
[project]
name = "data-science-tutorial"
version = "0.1.0"
description = "Data science tutorial RL environment using Jupyter notebook"
requires-python = ">=3.11"
dependencies = [
    "hud-python>=0.5.0",
    "openai>=1.0.0",
    "jupyter-client>=8.0.0",
    "jupyter-kernel-gateway>=3.0.0",
    "ipython>=8.0.0",
    "ipykernel>=6.0.0",
    "tornado>=6.0.0",
    "nbformat>=5.7.0",
    "numpy>=2.3.5",
    "scipy>=1.16.3",
    "pandas>=2.0.0",
    "anthropic>=0.30.0",
    "google-genai>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Specify the package directory (will be created in Dockerfile)
packages = ["data_science_tutorial"]
